# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for Osau AMS.",
        "title": "Osau AMS API",
        "contact": {
            "name": "API Support",
            "email": "bakit.german.work@icloud.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
...
        "/disciplines": {
            "get": {
                "description": "Get a list of all disciplines",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Disciplines"
                ],
                "summary": "Get all disciplines",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.DisciplineInfo"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing discipline",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Disciplines"
                ],
                "summary": "Update a discipline",
                "parameters": [
                    {
                        "description": "Discipline info",
                        "name": "discipline",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.PutDisciplineRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new discipline",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Disciplines"
                ],
                "summary": "Create a discipline",
                "parameters": [
                    {
                        "description": "Discipline info",
                        "name": "discipline",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateDisciplineRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Partially update an existing discipline",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Disciplines"
                ],
                "summary": "Partially update a discipline",
                "parameters": [
                    {
                        "description": "Discipline info",
                        "name": "discipline",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.PatchDisciplineRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/disciplines/departament/{departament_id}": {
            "get": {
                "description": "Get a list of disciplines by departament ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Disciplines"
                ],
                "summary": "Get disciplines by departament ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Departament ID",
                        "name": "departament_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.DisciplineInfo"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/disciplines/name/{name}": {
            "get": {
                "description": "Get a discipline by its name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Disciplines"
                ],
                "summary": "Get a discipline by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Discipline Name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.DisciplineInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/disciplines/{id}": {
            "get": {
                "description": "Get a discipline by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Disciplines"
                ],
                "summary": "Get a discipline by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Discipline ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.DisciplineInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an existing discipline",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Disciplines"
                ],
                "summary": "Delete a discipline",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Discipline ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Discipline": {
            "type": "object",
            "properties": {
                "departament_id": {
                    "type": "integer"
                },
                "discipline_id": {
                    "type": "integer"
                },
                "discipline_name": {
                    "type": "string"
                }
            }
        },
        "domain.DisciplineInfo": {
            "type": "object",
            "properties": {
                "discipline": {
                    "$ref": "#/definitions/domain.Discipline"
                },
                "discipline_sub": {
                    "$ref": "#/definitions/domain.DisciplineSub"
                }
            }
        },
        "domain.DisciplineSub": {
            "type": "object",
            "properties": {
                "departament_name": {
                    "type": "string"
                }
            }
        },
...
        "handler.CreateDisciplineRequest": {
            "type": "object",
            "required": [
                "departament_id",
                "discipline_name"
            ],
            "properties": {
                "departament_id": {
                    "type": "integer"
                },
                "discipline_name": {
                    "type": "string"
                }
            }
        },
...
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handler.PatchDisciplineRequest": {
            "type": "object",
            "required": [
                "discipline_id"
            ],
            "properties": {
                "departament_id": {
                    "type": "integer"
                },
                "discipline_id": {
                    "type": "integer"
                },
                "discipline_name": {
                    "type": "string"
                }
            }
        },
...
        "handler.PutDisciplineRequest": {
            "type": "object",
            "required": [
                "departament_id",
                "discipline_id",
                "discipline_name"
            ],
            "properties": {
                "departament_id": {
                    "type": "integer"
                },
                "discipline_id": {
                    "type": "integer"
                },
                "discipline_name": {
                    "type": "string"
                }
            }
        },
...
        "handler.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}